Chapter02/myblog/requirements.txt
Django
djangorestframework
psycopg2-binary
redis